requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.2.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
loguru==0.7.3
lxml==5.3.1
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
openai==0.28.0
packaging==24.2
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.1.0
pip==25.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pyzmq==26.3.0
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.45.1
XlsxWriter==3.2.2
yarl==1.18.3
zipp==3.21.0
